+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## Unreleased
+- Kotlin 1.6.10
+- Gradle 7.3.2
+- Coroutines 1.5.2
+- Added refreshVersions plugin
+- New kotlin-redux-compose module
 
 ## [0.5.5] - 2020-08-16
  - update to Kotlin 1.4.0